refactor: switch nav and header names
